isd_p - tweaked the width and height in photoshop to narrow it to the dimensions pictured.
grid.css - text-align: left; [for .col-1of2 and .col-1of3. the reason I didn't align left for 1/4 was that this was messing up the titles of the nav bar]

main.css - .item_description [gave i_d h2 a top margin of 1.5 em to align with the top of the left column photo.][adjusted width and height of the h2 font to affect scaling the image for credits]

Product_1.html - Added grid div, added two col-1of2s. Within the left column I placed the product title (h2) and the perspective image in custom dimensions to stretch to fit properly. Three more orthographic views added and set to the next level down in the column with heights and widths of 30% (accounting for ~1/3 of the column giving leeway for margins and padding.

-In the second column I placed an img tag within an h2 tag which links the credits image with the amount of money (defined relationship within main.css). A div with the add to cart button has an added strong tag to give the letters more weight on the reduced box dimensions. Added in the flavour text and an about to add an unordered list showing the technical specifications of each ship.