The work on the main CSS file revolves almost completely around the table,tr,&th styling sections to set up the look of the table.
The product_1.html changes were 2 attempts to create that table to display ship information. These 2 have been commented out for now and a third is created which looks good so far. Might end up staying in the right column or moving to the left column bottom if it looks better there.